Wait for informers
In runRootCommandWithProviderAndClient(), the pod and scm informers are
started before calling the provider init function. However, Start() is
asynchronous, since it calls informer.Run() via a goroutine. Thus there
is still a race condition, where the provider might get a
ResourceManager that is non-functional.

I added a WaitForCacheSync() call for both informers to ensure they are
operational before calling the provider init function.